feat: make a separate component
